3D Panning Complete
