+++ src/player/proto/audio-graph-player-abs-cont-tones.ts
@@ -46,7 +46,8 @@ import {
   getEndTime1,
   getStartTime1,
   glyphSorterByEnd,
-  glyphSorterByStart
+  glyphSorterByStart,
+  createPanner
 } from '../../util';
 import { AudioPrimitiveBuffer } from '../../pulse';
 import { AudioFilterPrototype } from '../../audioFilters';
@@ -127,27 +128,10 @@ export async function playAbsoluteContinuousTones(
   const gain = ctx.createGain();
   gain.connect(destination);
   
+  //DONE - function to handle this
   // decide between stereo or 3d pan
   const cartesianInputs = ['panX', 'panY', 'panZ'].filter(key => queue[0][key] !== undefined).length;
-  let panner!: AudioNode;
-
-  if (cartesianInputs == 1) {
-    const stereoPanner = ctx.createStereoPanner();
-    stereoPanner.connect(gain);
-    panner = stereoPanner;
-  } else {
-    const panner3D = ctx.createPanner();
-    panner3D.connect(gain);
-    panner3D.panningModel = 'HRTF';
-    panner3D.distanceModel = 'inverse';
-    panner3D.refDistance = 1;
-    panner3D.maxDistance = 10000;
-    panner3D.rolloffFactor = 1;
-    panner3D.coneInnerAngle = 360;
-    panner3D.coneOuterAngle = 0;
-    panner3D.coneOuterGain = 0;
-    panner = panner3D;
-  }
+  const panner = createPanner(ctx, cartesianInputs, gain);
 
   let sid = genRid()
   sendToneStartEvent({ sid });
@@ -199,11 +183,16 @@ export async function playAbsoluteContinuousTones(
       rampBy(sound.isFirst ? 'setValueAtTime' : rampers.loudness, gain.gain, sound.loudness, st);
     }
     // panner node
-    // TODO - Check stereo vs 3d
-    if (sound.pan !== undefined) {
-      // [check output:] panner.pan.setTargetAtTime(sound.pan, st, 0.35);
-      // rampBy(sound.isFirst ? 'setValueAtTime' : rampers.pan, panner.pan, sound.pan, st);
-      rampBy(sound.isFirst ? 'setTargetAtTime' : rampers.pan, panner.pan, sound.pan, st, 0.35);
+    // DONE - Check stereo vs 3d
+    const isStereo = cartesianInputs === 1 && queue[0].panX !== undefined && queue[0].panY === undefined && queue[0].panZ === undefined;
+    if (isStereo && sound.pan !== undefined) {
+      rampBy(sound.isFirst ? 'setTargetAtTime' : rampers.pan, (panner as StereoPannerNode).pan, sound.pan, st, 0.35);
+    } else if (!isStereo) {
+      if (sound.panX !== undefined && sound.panY !== undefined && sound.panZ !== undefined) {
+        (panner as PannerNode).positionX.setValueAtTime(sound.panX, st);
+        (panner as PannerNode).positionY.setValueAtTime(sound.panY, st);
+        (panner as PannerNode).positionZ.setValueAtTime(sound.panZ, st);
+      }
     }
 
     if (sound.isFirst) {
